reproify: consolidate to the real-processor OOM repro
Delete the hand-rolled OOM repro and rename the real-processor version into
its place (oversized-settlement-crash.{test,child}.ts). The real-processor
version is a superset in coverage — it exercises the same journaling +
storage + classification the hand-rolled one did, plus the genuine
AgentProcessor/CapabilityHostProcessor folds — so keeping both just
double-pinned one bug. One repro, canonical name.
